Feature android/event api

See merge request s10-webmobile4-sub2/S10P12A301!210

+++ Android/TogeDuck/app/src/main/java/com/idle/togeduck/common/model/DefaultResponse.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class DefaultResponse (
-    val code: Int,
+    val code: Long,
     val message: String
 )
